[diffusion, rollout, tests] refactor: consolidate rollout postprocessor wrapper into diffusion_rollout_output
Move the postprocessor wrapper logic into `diffusion_rollout_output.py` as
`wrap_rollout_postprocessor`, remove `vllm_omni_output_compat.py`, and update
rollout adapters and tests.
