+++ verl_omni/pipelines/wan22_dance_grpo/vllm_omni_rollout_adapter.py
@@ -36,10 +36,12 @@ from vllm_omni.diffusion.models.wan2_2.pipeline_wan2_2 import Wan22Pipeline, ret
 from vllm_omni.diffusion.request import OmniDiffusionRequest
 from vllm_omni.platforms import current_omni_platform
 
-from verl_omni.pipelines.diffusion_rollout_output import rollout_output
+from verl_omni.pipelines.diffusion_rollout_output import (
+    rollout_output,
+    wrap_rollout_postprocessor,
+)
 from verl_omni.pipelines.model_base import VllmOmniPipelineBase
 from verl_omni.pipelines.schedulers import FlowMatchSDEDiscreteScheduler
-from verl_omni.pipelines.vllm_omni_output_compat import envelope_aware_postprocessor
 
 from .common import sd3_time_shift, seed_from_prompt_ids
 
@@ -51,7 +53,7 @@ _WAN_POST_PROCESS_FACTORY = pipeline_wan2_2.get_wan22_post_process_func
 
 def get_rollout_post_process_func(od_config):
     """Postprocess Wan media while preserving rollout metadata."""
-    return envelope_aware_postprocessor(_WAN_POST_PROCESS_FACTORY(od_config))
+    return wrap_rollout_postprocessor(_WAN_POST_PROCESS_FACTORY(od_config))
 
 
 # vllm-omni resolves the built-in architecture's factory in the engine process.
